+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - Server: API was renamed to server.
+- Server: The `--initialize-database` flag was changed to the `initialize` subcommand.
 
 ## [0.2.0] - 2018-11-02